Move either instances
