+++ core/src/main/scala/au/id/tmm/bfect/instances/package.scala
@@ -1,0 +1,7 @@
+package au.id.tmm.bfect
+
+package object instances {
+  object either extends EitherInstances
+
+  object all extends AnyRef with EitherInstances
+}
