+++ core/src/test/scala/au/id/tmm/bfect/BifunctorMonadSpec.scala
@@ -15,13 +15,12 @@
   */
 package au.id.tmm.bfect
 
-import au.id.tmm.bfect.syntax.bifunctorMonad._
 import au.id.tmm.bfect.BifunctorMonadSpec._
+import au.id.tmm.bfect.instances.either._
+import au.id.tmm.bfect.syntax.bifunctorMonad._
 import org.scalatest.flatspec.AnyFlatSpec
 
 class BifunctorMonadSpec extends AnyFlatSpec {
-
-  import EitherInstances.biFunctorMonadInstance
 
   "absolve" should "absolve when the outer error type is Nothing" in {
     val nestedEither: Either[Nothing, Either[ChildError, Unit]] = Right(Right(()))
